taipei-day-trip/requirements.txt
fastapi
mysql-connector-python
uvicorn
pydantic
python-dotenv
python-multipart
pyjwt
requests
passlib[bcrypt]
import httpx